add open positions list and trade closing functionality

+++ src/main/java/com/example/ea_beadando/model/forex/MessageClosePosition.java
@@ -1,8 +1,8 @@
 package com.example.ea_beadando.model.forex;
 
 public class MessageClosePosition {
-    private int tradeId;
+    private long tradeId;
     // Getters/Setters
-    public int getTradeId() { return tradeId; }
+    public int getTradeId() { return Math.toIntExact(tradeId); }
     public void setTradeId(int tradeId) { this.tradeId = tradeId; }
 }
